+++ Sources/StreamChatSwiftUI/ChatChannel/ChannelInfo/MediaAttachmentsViewModel.swift
@@ -19,8 +19,8 @@ class MediaAttachmentsViewModel: ObservableObject, ChatMessageSearchControllerDe
 
     private var loadingNextMessages = false
 
-    var allImageAttachments: [ChatMessageImageAttachment] {
-        mediaItems.compactMap(\.imageAttachment)
+    var allMediaAttachments: [MediaAttachment] {
+        mediaItems.compactMap(\.mediaAttachment)
     }
 
     init(channel: ChatChannel) {
@@ -106,11 +106,34 @@ class MediaAttachmentsViewModel: ObservableObject, ChatMessageSearchControllerDe
     }
 }
 
-struct MediaItem: Identifiable {
-    let id: String
-    let isVideo: Bool
-    let message: ChatMessage
+public struct MediaItem: Identifiable {
+    public let id: String
+    public let isVideo: Bool
+    public let message: ChatMessage
 
-    var videoAttachment: ChatMessageVideoAttachment?
-    var imageAttachment: ChatMessageImageAttachment?
+    public var videoAttachment: ChatMessageVideoAttachment?
+    public var imageAttachment: ChatMessageImageAttachment?
+    
+    public init(
+        id: String,
+        isVideo: Bool,
+        message: ChatMessage,
+        videoAttachment: ChatMessageVideoAttachment?,
+        imageAttachment: ChatMessageImageAttachment?
+    ) {
+        self.id = id
+        self.isVideo = isVideo
+        self.message = message
+        self.videoAttachment = videoAttachment
+        self.imageAttachment = imageAttachment
+    }
+    
+    public var mediaAttachment: MediaAttachment? {
+        if let videoAttachment {
+            return MediaAttachment(url: videoAttachment.videoURL, type: .video)
+        } else if let imageAttachment {
+            return MediaAttachment(url: imageAttachment.imageURL, type: .image)
+        }
+        return nil
+    }
 }